fix: account generation details
Fixes #135
