+++ swarm-ui/src/lib/components/drawer.svelte
@@ -22,6 +22,8 @@
   import Rocket from 'carbon-icons-svelte/lib/Rocket.svelte'
   import TrashCan from 'carbon-icons-svelte/lib/TrashCan.svelte'
   import Logout from 'carbon-icons-svelte/lib/Logout.svelte'
+  import View from 'carbon-icons-svelte/lib/View.svelte'
+  import ViewOff from 'carbon-icons-svelte/lib/ViewOff.svelte'
   import NetworkSettingsModal from './network-settings-modal.svelte'
   import ThemeToggle from './theme-toggle.svelte'
   import FlexItem from '$lib/components/ui/flex-item.svelte'
@@ -33,13 +35,15 @@
   import { connectedAppsStore } from '$lib/stores/connected-apps.svelte'
   import { postageStampsStore } from '$lib/stores/postage-stamps.svelte'
   import type { Account, Identity } from '$lib/types'
+  import type { Bytes } from '@ethersphere/bee-js'
+  import { deriveSecretSeedEncryptionKey, decryptSecretSeed } from '$lib/utils/encryption'
+  import { getMasterKeyFromAccount } from '$lib/utils/account-auth'
   import EthereumLogo from './ethereum-logo.svelte'
   import PasskeyLogo from './passkey-logo.svelte'
   import Input from './ui/input/input.svelte'
   import Typography from './ui/typography.svelte'
   import CopyButton from './copy-button.svelte'
   import Tooltip from './ui/tooltip.svelte'
-  import ViewGenerationDetailsModal from './view-generation-details-modal.svelte'
   import DeleteModal from './delete-modal.svelte'
   import Bot from 'carbon-icons-svelte/lib/Bot.svelte'
 
@@ -52,14 +56,20 @@
 
   let { drawerOpen = $bindable(), identities, account, identityId }: Props = $props()
 
-  let screen = $state<'main' | 'all-accounts' | 'account-details'>('main')
+  let screen = $state<'main' | 'all-accounts' | 'account-details' | 'generation-details'>('main')
   // eslint-disable-next-line svelte/prefer-writable-derived
   let accountName = $state('')
   let showUpgradeTooltip = $state(false)
   let showExportTooltip = $state(false)
   let networkSettingsModalOpen = $state(false)
-  let showGenerationDetailsModal = $state(false)
   let showDeleteModal = $state(false)
+
+  // Generation details state
+  let isUnmasked = $state(false)
+  let secretSeed = $state('')
+  let generationDetailsError = $state<string | undefined>(undefined)
+  let isAuthenticating = $state(false)
+  let decryptedMasterKey: Bytes | undefined
 
   $effect(() => {
     accountName = account.name
@@ -173,6 +183,51 @@
     sessionStore.clearAccount()
     drawerOpen = false
     await goto(resolve(routes.HOME))
+  }
+
+  async function handleAuthenticate() {
+    if (account.type !== 'ethereum') return
+
+    try {
+      isAuthenticating = true
+      generationDetailsError = undefined
+      decryptedMasterKey = await getMasterKeyFromAccount(account)
+    } catch (err) {
+      generationDetailsError = err instanceof Error ? err.message : 'Authentication failed'
+      console.error('Authentication failed:', err)
+    } finally {
+      isAuthenticating = false
+    }
+  }
+
+  async function handleUnmask() {
+    if (!decryptedMasterKey) {
+      await handleAuthenticate()
+    }
+
+    try {
+      if (decryptedMasterKey && account.type === 'ethereum') {
+        const secretSeedEncryptionKey = await deriveSecretSeedEncryptionKey(decryptedMasterKey)
+        secretSeed = await decryptSecretSeed(account.encryptedSecretSeed, secretSeedEncryptionKey)
+        isUnmasked = true
+      }
+    } catch (err) {
+      generationDetailsError = err instanceof Error ? err.message : 'Failed to decrypt secret seed'
+      console.error('Failed to decrypt secret seed:', err)
+    }
+  }
+
+  function handleMask() {
+    isUnmasked = false
+    secretSeed = ''
+  }
+
+  function handleGenerationDetailsClose() {
+    isUnmasked = false
+    secretSeed = ''
+    decryptedMasterKey = undefined
+    generationDetailsError = undefined
+    screen = 'account-details'
   }
 </script>
 
@@ -434,9 +489,7 @@
             <Button
               variant="ghost"
               dimension="compact"
-              onclick={() => {
-                showGenerationDetailsModal = true
-              }}
+              onclick={() => (screen = 'generation-details')}
               leftAlign
             >
               <IbmCloudHyperProtectCryptoServices size={20} />
@@ -552,19 +605,94 @@
             Delete account
           </Button>
         </Vertical>
+      {:else if screen === 'generation-details'}
+        <Horizontal
+          --horizontal-gap="var(--double-padding)"
+          --horizontal-justify-content="space-between"
+          --horizontal-align-items="center"
+          style="padding: var(--padding)"
+        >
+          <Horizontal --horizontal-gap="var(--half-padding)">
+            <Button variant="ghost" dimension="compact" onclick={() => (screen = 'account-details')}
+              ><ChevronLeft size={20} /></Button
+            >
+            Account generation details
+          </Horizontal>
+          <Button
+            variant="ghost"
+            dimension="compact"
+            onclick={() => {
+              handleGenerationDetailsClose()
+              drawerOpen = false
+            }}><CloseLarge size={20} /></Button
+          >
+        </Horizontal>
+        <Vertical --vertical-gap="var(--padding)" style="padding: 0 var(--padding)">
+          <Horizontal --horizontal-gap="var(--half-padding)" --horizontal-align-items="end">
+            <Input
+              variant="outline"
+              dimension="compact"
+              value={account.type === 'ethereum' ? account.ethereumAddress.toHex() : ''}
+              class="grower"
+              label="Initial wallet address"
+              disabled
+            />
+            <div style="border: 1px solid transparent">
+              <CopyButton
+                text={account.type === 'ethereum' ? account.ethereumAddress.toHex() : ''}
+              />
+            </div>
+          </Horizontal>
+          <Horizontal --horizontal-gap="var(--half-padding)" --horizontal-align-items="end">
+            <Input
+              variant="outline"
+              dimension="compact"
+              value={isUnmasked ? secretSeed : '••••••••'}
+              class="grower"
+              label="Secret seed"
+              disabled
+              type={isUnmasked ? 'text' : 'password'}
+            />
+            <div style="border: 1px solid transparent">
+              {#if !isUnmasked}
+                <Button
+                  dimension="compact"
+                  variant="ghost"
+                  onclick={handleUnmask}
+                  disabled={isAuthenticating}
+                  title="Unmask secret seed"
+                >
+                  <View size={20} />
+                </Button>
+              {:else}
+                <Button
+                  dimension="compact"
+                  variant="ghost"
+                  onclick={handleMask}
+                  title="Mask secret seed"
+                >
+                  <ViewOff size={20} />
+                </Button>
+              {/if}
+            </div>
+          </Horizontal>
+          <Typography>
+            This secret seed is used in combination with your wallet to restore your Swarm ID
+            account. <b
+              >Store it in a password manager or write it down on a piece of paper hidden in a
+              secure location. Never disclose it to anyone.</b
+            >
+          </Typography>
+          {#if generationDetailsError}
+            <Typography style="color: var(--colors-red)"
+              >There was an error during authentication</Typography
+            >
+          {/if}
+        </Vertical>
       {/if}
     </Vertical>
   </div>
 </div>
-{#if account.type === 'ethereum'}
-  <ViewGenerationDetailsModal
-    bind:open={showGenerationDetailsModal}
-    bind:drawerOpen
-    {account}
-    onClose={() => (showGenerationDetailsModal = false)}
-  />
-{/if}
-
 <NetworkSettingsModal bind:open={networkSettingsModalOpen} />
 
 <DeleteModal
